requirements.txt
django==4.2
djangorestframework==3.14.0
django-environ==0.10.0
psycopg2-binary==2.9.6
